+++ BodyTemplatesTableView.m
@@ -38,7 +38,7 @@ static NSString* BodyTemplatesPboardType = @"BodyTemplatesPboardType"; //pboard 
 -(BOOL) acceptsFirstMouse:(NSEvent *)theEvent //using the tableview does not need to activate the window first
 {
   NSPoint point = [self convertPoint:[theEvent locationInWindow] fromView:nil];
-  int row = [self rowAtPoint:point];
+  NSInteger row = [self rowAtPoint:point];
   [self selectRowIndexes:[NSIndexSet indexSetWithIndex:row] byExtendingSelection:NO];
   return YES;
 }
@@ -55,7 +55,7 @@ static NSString* BodyTemplatesPboardType = @"BodyTemplatesPboardType"; //pboard 
 //edit selected row
 -(IBAction) edit:(id)sender
 {
-  int selectedRow = [self selectedRow];
+  NSInteger selectedRow = [self selectedRow];
   if (selectedRow >= 0)
     [self editColumn:0 row:selectedRow withEvent:nil select:YES];
 }
