+++ TextShortcutsTableView.m
@@ -50,7 +50,7 @@ static NSString* EditionTextShortcutsPboardType = @"EditionTextShortcutsPboardTy
 -(BOOL) acceptsFirstMouse:(NSEvent *)theEvent //using the tableview does not need to activate the window first
 {
   NSPoint point = [self convertPoint:[theEvent locationInWindow] fromView:nil];
-  int row = [self rowAtPoint:point];
+  NSInteger row = [self rowAtPoint:point];
   [self selectRowIndexes:[NSIndexSet indexSetWithIndex:row] byExtendingSelection:NO];
   return YES;
 }
@@ -67,7 +67,7 @@ static NSString* EditionTextShortcutsPboardType = @"EditionTextShortcutsPboardTy
 //edit selected row
 -(IBAction) edit:(id)sender
 {
-  int selectedRow = [self selectedRow];
+  NSInteger selectedRow = [self selectedRow];
   if (selectedRow >= 0)
     [self editColumn:0 row:selectedRow withEvent:nil select:YES];
 }
@@ -111,7 +111,7 @@ static NSString* EditionTextShortcutsPboardType = @"EditionTextShortcutsPboardTy
 
 -(void) moveUp:(id)sender
 {
-  int selectedRow = [self selectedRow];
+  NSInteger selectedRow = [self selectedRow];
   if (selectedRow > 0)
     --selectedRow;
   [self selectRowIndexes:[NSIndexSet indexSetWithIndex:selectedRow] byExtendingSelection:NO];
@@ -121,7 +121,7 @@ static NSString* EditionTextShortcutsPboardType = @"EditionTextShortcutsPboardTy
 
 -(void) moveDown:(id)sender
 {
-  int selectedRow = [self selectedRow];
+  NSInteger selectedRow = [self selectedRow];
   if ((selectedRow >= 0) && (selectedRow+1 < [self numberOfRows]))
     ++selectedRow;
   [self selectRowIndexes:[NSIndexSet indexSetWithIndex:selectedRow] byExtendingSelection:NO];
@@ -132,7 +132,7 @@ static NSString* EditionTextShortcutsPboardType = @"EditionTextShortcutsPboardTy
 //prevents from selecting next line when finished editing
 -(void) textDidEndEditing:(NSNotification *)aNotification
 {
-  int selectedRow = [self selectedRow];
+  NSInteger selectedRow = [self selectedRow];
   [super textDidEndEditing:aNotification];
   [self selectRowIndexes:[NSIndexSet indexSetWithIndex:selectedRow] byExtendingSelection:NO];
 }
@@ -141,7 +141,7 @@ static NSString* EditionTextShortcutsPboardType = @"EditionTextShortcutsPboardTy
 //delegate methods
 -(void) tableViewSelectionDidChange:(NSNotification *)aNotification
 {
-  unsigned int lastIndex = [[self selectedRowIndexes] lastIndex];
+  NSUInteger lastIndex = [[self selectedRowIndexes] lastIndex];
   [self scrollRowToVisible:lastIndex];
 }
 //end tableViewSelectionDidChange:
